+++ src/main/webapp/resources/js/pages/pipelines/pipeline.launch.js
@@ -1,5 +1,6 @@
 import "jquery";
 import angular from "angular";
+import "angular-ui-bootstrap";
 import "ng-file-upload";
 
 import React from "react";
@@ -656,7 +657,7 @@ function FileUploadCtrl($rootScope, Upload) {
 }
 
 const pipelineModule = angular
-  .module("irida.pipelines", ["irida.cart", "ngFileUpload"])
+  .module("irida.pipelines", ["ui.bootstrap", "irida.cart", "ngFileUpload"])
   .controller("PipelineController", [
     "$rootScope",
     "$http",